Introducing a global garbage collection map for handling Rados operations which are not valid anymore.
It might happen that some operations are cancelled due to
user requests or system errors. However, Rados is asynchronous,
so it may happen that some of the operations are running and
they need to be cleaned up properly as each Rados Op
has its own UserCtx allocated (need to deallocate the memory).